+++ test/test_parse.c
@@ -3249,6 +3249,12 @@ START_TEST(test_issue367)
 {
    input_from_file(TESTDIR "/parse/issue367.vhd");
 
+   const error_t expect[] = {
+      { 13, "declaration of I hides an earlier declaration with the same " },
+      { -1, NULL }
+   };
+   expect_errors(expect);
+
    tree_t p = parse();
    fail_if(p == NULL);
    fail_unless(tree_kind(p) == T_PACKAGE);
@@ -3278,7 +3284,7 @@ START_TEST(test_issue367)
 
    fail_if(parse() != NULL);
 
-   fail_if_errors();
+   check_expected_errors();
 }
 END_TEST
 
@@ -6157,6 +6163,12 @@ START_TEST(test_issue870)
 
    input_from_file(TESTDIR "/parse/issue870.vhd");
 
+   const error_t expect[] = {
+      { 13, "declaration of SEED1 hides an earlier declaration with the " },
+      { -1, NULL }
+   };
+   expect_errors(expect);
+
    tree_t e = parse();
    fail_if(e == NULL);
    fail_unless(tree_kind(e) == T_ENTITY);
@@ -6168,7 +6180,7 @@ START_TEST(test_issue870)
 
    fail_unless(parse() == NULL);
 
-   fail_if_errors();
+   check_expected_errors();
 }
 END_TEST
 
@@ -6661,6 +6673,32 @@ START_TEST(test_issue917)
 }
 END_TEST
 
+START_TEST(test_issue905)
+{
+   input_from_file(TESTDIR "/parse/issue905.vhd");
+
+   const error_t expect[] = {
+      { 36, "declaration of I hides an earlier declaration with the "
+        "same type" },
+      { -1, NULL }
+   };
+   expect_errors(expect);
+
+   tree_t e = parse();
+   fail_if(e == NULL);
+   fail_unless(tree_kind(e) == T_ENTITY);
+   lib_put(lib_work(), e);
+
+   tree_t a = parse();
+   fail_if(a == NULL);
+   fail_unless(tree_kind(a) == T_ARCH);
+
+   fail_unless(parse() == NULL);
+
+   check_expected_errors();
+}
+END_TEST
+
 Suite *get_parse_tests(void)
 {
    Suite *s = suite_create("parse");
@@ -6815,6 +6853,7 @@ Suite *get_parse_tests(void)
    tcase_add_test(tc_core, test_defer1);
    tcase_add_test(tc_core, test_error13);
    tcase_add_test(tc_core, test_issue917);
+   tcase_add_test(tc_core, test_issue905);
    suite_add_tcase(s, tc_core);
 
    return s;